infrastructure changes
removing "/server/conf" route. we do not want to print configuration with password and such.
adding spring. we also have guice for jcloud at the moment - I hope we can convert it to spring as well.
adding redirects to "/" - checks if there is an authKey in the cookie. If so, redirect to "/admin/widgets.html" otherwise, to signin
refactoring configuration. We should use play 2.0 awesome configuration features.